Moved most src files to adaptlil.
