+++ src/adaptlil/data_classes/Blink.java
@@ -1,0 +1,4 @@
+package adaptlil.data_classes;
+
+public class Blink {
+}
