+++ src/adaptlil/data_classes/RightEyePointOfGaze.java
@@ -1,49 +1,46 @@
-package data_classes;
+package adaptlil.data_classes;
 
 import adaptlil.annotations.IgnoreWekaAttribute;
-import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlProperty;
-import interpolation.Interpolation;
-import adaptlil.serialization_helpers.IntToBooleanDeserializer;
+import adaptlil.interpolation.Interpolation;
 
-public class LeftEyePointOfGaze {
-
+public class RightEyePointOfGaze {
 
     //Default constructor for jackson
-    public LeftEyePointOfGaze() {}
+    public RightEyePointOfGaze() {}
 
     /**
      *
-     * @param x The X-coordinate of the left eye POG, as a fraction of the screen size.
-     * @param y The Y-coordinate of the left eye POG, as a fraction of the screen size.
+     * @param x The X-coordinate of the right eye POG, as a fraction of the screen size.
+     * @param y The Y-coordinate of the right eye POG, as a fraction of the screen size.
      * @param isValid Flag that details if the tracker data is valid or not
      */
-    public LeftEyePointOfGaze(double x, double y, boolean isValid) {
+    public RightEyePointOfGaze(double x, double y, boolean isValid) {
         this.x = x;
         this.y = y;
         this.isValid = isValid;
     }
 
-    @JacksonXmlProperty(isAttribute = true, localName = "LPOGX")
+    @JacksonXmlProperty(isAttribute = true, localName = "RPOGX")
     private double x;
-    @JacksonXmlProperty(isAttribute = true, localName = "LPOGY")
+    @JacksonXmlProperty(isAttribute = true, localName = "RPOGY")
     private double y;
+
     @IgnoreWekaAttribute
-    @JsonDeserialize(using = IntToBooleanDeserializer.class)
-    @JacksonXmlProperty(isAttribute = true, localName = "LPOGV")
+    @JacksonXmlProperty(isAttribute = true, localName = "RPOGV")
     private boolean isValid;
 
     /**
-     * The X-coordinate of the left eye POG, as a fraction of the screen size.
-     * @param x X-coordinate of the left eye POG, as a fraction of the screen size.
+     * The X-coordinate of the right eye POG, as a fraction of the screen size.
+     * @param x X-coordinate of the right eye POG, as a fraction of the screen size.
      */
     public void setX(double x) {
         this.x = x;
     }
 
     /**
-     * The Y-coordinate of the left eye POG, as a fraction of the screen size.
-     * @param y Y-coordinate of the left eye POG, as a fraction of the screen size.
+     * The Y-coordinate of the right eye POG, as a fraction of the screen size.
+     * @param y Y-coordinate of the right eye POG, as a fraction of the screen size.
      */
 
     public void setY(double y) {
@@ -59,7 +56,7 @@ public class LeftEyePointOfGaze {
     }
 
     /**
-     * The X-coordinate of the left eye POG, as a fraction of the screen size.
+     * The X-coordinate of the right eye POG, as a fraction of the screen size.
      * @return double
      */
     public double getX() {
@@ -67,7 +64,7 @@ public class LeftEyePointOfGaze {
     }
 
     /**
-     * The Y-coordinate of the left eye POG, as a fraction of the screen size.
+     * The Y-coordinate of the right eye POG, as a fraction of the screen size.
      * @return double
      */
     public double getY() {
@@ -88,20 +85,20 @@ public class LeftEyePointOfGaze {
      * @param steps
      * @return
      */
-    public LeftEyePointOfGaze[] interpolate(LeftEyePointOfGaze a, LeftEyePointOfGaze b, int steps) {
+    public RightEyePointOfGaze[] interpolate(RightEyePointOfGaze a, RightEyePointOfGaze b, int steps) {
         Interpolation interpolation = new Interpolation();
-        LeftEyePointOfGaze[] leftEyeInterpols = new LeftEyePointOfGaze[steps];
+        RightEyePointOfGaze[] rightEyeInterpols = new RightEyePointOfGaze[steps];
         double[] aCoords = new double[]{a.getX(), a.getY()};
         double[] bCoords = new double[]{b.getX(), b.getY()};
         double[][] interpolCoords = interpolation.interpolate(aCoords, bCoords, steps);
 
         for (int i = 0; i < steps; ++i) {
-            LeftEyePointOfGaze c = new LeftEyePointOfGaze();
+            RightEyePointOfGaze c = new RightEyePointOfGaze();
             c.setX(interpolCoords[i][0]);
             c.setY(interpolCoords[i][1]);
             c.setValid(true);
-            leftEyeInterpols[i] = c;
+            rightEyeInterpols[i] = c;
         }
-        return leftEyeInterpols;
+        return rightEyeInterpols;
     }
 }
