+++ src/adaptlil/data_classes/geometry/Cartesian2D.java
@@ -1,4 +1,4 @@
-package geometry;
+package adaptlil.data_classes.geometry;
 
 public class Cartesian2D {
     protected float x;
