+++ src/adaptlil/data_classes/geometry/Vector2D.java
@@ -1,4 +1,4 @@
-package geometry;
+package adaptlil.data_classes.geometry;
 
 //Currently no use but I want to have it here just in case...
 public class Vector2D {
