+++ src/adaptlil/gazepoint/excel/GazeClassificationData.java
@@ -1,4 +1,4 @@
-package ontomap;
+package adaptlil.gazepoint.excel;
 
 import adaptlil.GazeWindow;
