--- src/data_classes/Blink.java
@@ -1,4 +1,0 @@
-package data_classes;
-
-public class Blink {
-}
